requirements.txt
numpy==1.16.4
tensorflow-gpu==1.4.0
h5py==2.8.0
matplotlib==2.2.3
open3d-python==0.3.0
scikit-learn==0.20.3
scipy==1.2.1
argparse==1.1
